have google maps showing map and finding tour from GoogleId
